more cleanup, removing backfill -- unused
